New Classes added for posting status with media(image/video) on a User Page and storing them into MongoDB.
A new package is created with classes for practising Facebook GET requests using RestFB library.

+++ src/main/java/com/moves/movesCelebrity/social/models/FBPost.java
@@ -4,11 +4,11 @@ public class FBPost {
 
     String postId;
     String postContent;
-    Integer likesCount;
-    Integer shareCount;
-    Integer commentsCount;
+    Long likesCount;
+    Long shareCount;
+    Long commentsCount;
 
-    public FBPost(String postId, String postContent, Integer likesCount, Integer shareCount, Integer commentsCount) {
+    public FBPost(String postId, String postContent, Long likesCount, Long shareCount, Long commentsCount) {
         this.postId = postId;
         this.postContent = postContent;
         this.likesCount = likesCount;
@@ -35,27 +35,27 @@ public class FBPost {
         this.postContent = postContent;
     }
 
-    public Integer getLikesCount() {
+    public Long getLikesCount() {
         return likesCount;
     }
 
-    public void setLikesCount(Integer likesCount) {
+    public void setLikesCount(Long likesCount) {
         this.likesCount = likesCount;
     }
 
-    public Integer getShareCount() {
+    public Long getShareCount() {
         return shareCount;
     }
 
-    public void setShareCount(Integer shareCount) {
+    public void setShareCount(Long shareCount) {
         this.shareCount = shareCount;
     }
 
-    public Integer getCommentsCount() {
+    public Long getCommentsCount() {
         return commentsCount;
     }
 
-    public void setCommentsCount(Integer commentsCount) {
+    public void setCommentsCount(Long commentsCount) {
         this.commentsCount = commentsCount;
     }
 }
